requirements.txt
ruamel.yaml<0.15
simplejson
appdirs
requests
jenkinsapi
xmltodict
graphviz